Final bottles bonus fixes, hourglass timer hud padding, and airscore hud positioning
